Refactor retriever/websearch & improve speed/prompts
